# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: > 
    The Mechanic Shop API allows clients to manage customers, mechanics, service tickets, inventory, and related operations.
    It provides RESTful endpoints to create, retrieve, update, and delete records, supporting functionality such as
    assigning mechanics to service tickets, updating inventory, assigning parts to service tickets, and managing customer profiles.
    This API is designed to streamline the workflow of a mechanic shop and integrate with frontend or mobile interfaces.
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths: 
...
    /customers:
      # Get all customers
      get:
        tags:
          - "Customers"
        summary: "Get all customers"
        description: "Retrieve a list of all customers."
        responses:
          200:
            description: "A list of customers."
            schema:
              type: array
              items:
                $ref: '#/definitions/Customer'
      
      # Create a new customer
      post:
        tags:
          - "Customers"
        summary: "Create a new customer"
        description: "Add a new customer to the system."
        parameters:
          - in: body
            name: customer
            required: true
            schema:
              $ref: '#/definitions/CustomerInput'
        responses:
          201:
            description: "Customer created successfully."
            schema:
              $ref: '#/definitions/Customer'

    # Get, update, and delete a specific customer by ID
    /customers/{customer_id}:
      # Get a specific customer by ID
      get:
        tags:
          - "Customers"
        summary: "Get a customer by ID"
        description: "Retrieve a specific customer by their ID."
        parameters:
          - name: customer_id
            in: path
            required: true
            type: integer
        responses:
          200:
            description: "Customer details."
            schema:
              $ref: '#/definitions/Customer'
          404:
            description: "Customer not found."
    
      # Update a specific customer by ID
      put:
        tags:
          - "Customers"
        summary: "Update a customer"
        description: "Update an existing customer's information. Email can not be updated due to being used as a unique identifier for the customer."
        security:
          - BearerAuth: []
        parameters:
          - name: customer_id
            in: path
            required: true
            type: integer
          - in: body
            name: customer
            required: true
            schema:
              $ref: '#/definitions/CustomerUpdateInput'
        responses:
          200:
            description: "Customer updated successfully."
            schema:
              $ref: '#/definitions/Customer'
          404:
            description: "Customer not found."
    
      # Delete a specific customer by ID
      delete:
        tags:
          - "Customers"
        summary: "Delete a customer"
        description: "Remove a customer from the system."
        security:
          - BearerAuth: []
        parameters:
          - name: customer_id
            in: path
            required: true
            type: integer
        responses:
          204:
            description: "Customer deleted successfully."
          403:
            description: "Unauthorized"
          404:
            description: "Customer not found."

    # ------------------------ Mechanics -----------------------
    # Get all mechanics and create a new mechanic
...
    /inventory:
      # Get all inventory products
      get:
        tags:
          - "Inventory"
        summary: "Get all inventory products"
        description: "Retrieve a list of all products in the inventory."
        responses:
          200:
            description: "A list of inventory products."
            schema:
              type: array
              items:
                $ref: '#/definitions/InventoryProduct'
      
      # Create a new inventory product
      post:
        tags:
          - "Inventory"
        summary: "Create a new inventory product"
        description: "Add a new product to the inventory."
        parameters:
          - in: body
            name: InventoryItem
            required: true
            schema:
              $ref: '#/definitions/InventoryInput'
        responses:
          201:
            description: "Inventory Product created successfully."
            schema:
              $ref: '#/definitions/InventoryProduct'
    
    # Get, update, and delete a specific Inventory Product by ID
    /inventory/{inventory_id}:
      # Get a specific inventory product by ID
      get:
        tags:
          - "Inventory"
        summary: "Get a Inventory Product by ID"
        description: "Retrieve a specific inventory product by its ID."
        parameters:
          - name: inventory_id
            in: path
            required: true
            type: integer
        responses:
          200:
            description: "Inventory Product details."
            schema:
              $ref: '#/definitions/InventoryProduct'
          404:
            description: "Inventory Product not found."
      
      # Update a specific Inventory Product by ID
      put:
        tags:
          - "Inventory"
        summary: "Update a Inventory Product"
        description: "Update an existing Inventory Product's information."
        parameters:
          - name: inventory_id
            in: path
            required: true
            type: integer
          - in: body
            name: inventory_product
            required: true
            schema:
              $ref: '#/definitions/InventoryInput'
        responses:
          200:
            description: "Inventory Product updated successfully."
            schema:
              $ref: '#/definitions/InventoryProduct'
          404:
            description: "Inventory Product not found."
      
      # Delete a specific Inventory Product by ID
      delete:
        tags:
          - "Inventory"
        summary: "Delete a Inventory Product"
        description: "Remove a Inventory Product from the inventory."
        parameters:
          - name: inventory_id
            in: path
            required: true
            type: integer
        responses:
          204:
            description: "Inventory Product deleted successfully."
          404:
            description: "Inventory Product not found."

    # ------------------------ Service Tickets -----------------------
    # Get all service tickets and create a new service ticket
...
securityDefinitions:
    BearerAuth:
      type: "apiKey"
      name: "Authorization"
      in: "header"
      description: "Bearer token for authentication. You must log in via the /auth/login to receive a token. Example: 'Bearer your.token.here'"
definitions:
...
  CustomerInput:
    type: object
    required:
      - name
      - phone
      - password
    properties:
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        example: "john.customer@example.com"
      phone:
        type: string
        example: "555-555-5555"
      password:
        type: string
        example: "password123"
  CustomerUpdateInput:
    type: object
    required:
      - name
      - phone
      - password
    properties:
      name:
        type: string
        example: "John Doe"
      phone:
        type: string
        example: "555-555-5555"
      password:
        type: string
        example: "password123"
  Customer:
    type: object
    required:
      - id
      - name
      - email
      - phone
      - password
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        example: "john.customer@example.com"
      phone_number:
        type: string
        example: "555-555-5555"

# ------------------------ Mechanics -----------------------
...
  InventoryInput:
    type: object
    required:
      - name
      - price
    properties:
      name:
        type: string
        description: "Name of the part"
        example: "Brake Pad"
      price:
        type: number
        format: float
        description: "Price of the part"
        example: 29.99
  InventoryProduct:
    type: object
    required:
      - id
      - name
      - price
    properties:
      id:
        type: integer
        description: "Unique identifier for the part"
        example: 1
      name:
        type: string
        description: "Name of the part"
        example: "Brake Pad"
      price:
        type: number
        format: float
        description: "Price of the part"
        example: 29.99

# ------------------------ Service Tickets -----------------------
...